Removed borderType, moved InputBorder et al to its own file
